requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
xhtml2pdf>=0.2.11
Pillow>=10.0
